requirements.txt
flask==2.2.2
flask_socketio==5.3.2
python-dotenv==0.21.0